refactor: VRow is a single functional component
